feat: fix typo on jigglin_delay, now delay and delta work

+++ app/src-tauri/src/riggler_config.rs
@@ -3,7 +3,7 @@ use serde::{Deserialize, Serialize};
 #[derive(Debug, Serialize, Deserialize, Clone, Copy)]
 pub struct RigglerConfig {
     pub jiggling_delta: i32,
-    pub jiggling_delay: u64,
+    pub jiggling_delay: i32,
 }
 
 impl Default for RigglerConfig {
